migrate MessageManager to new file and add Disable_All option
